Move giving season header into its own component
